🔨 #73 코멘트 컴포넌트 세분화
